extract some type
