+++ modules/standards/decls/index.d.ts
@@ -1,4 +1,15 @@
-import type {
-} from 'phenyl-interfaces'
+import {
+  LoginCommand,
+  Session,
+  GeneralRequestData,
+  GeneralResponseData,
+} from '@phenyl/interfaces'
 
-export type EncryptFunction = (str: string, options?: Object) => string
+// @TODO: should we put those types into @phenyl/interfaces?
+type RestApiExecution = (
+    reqData: GeneralRequestData,
+    session: Session | null | undefined
+) => Promise<GeneralResponseData>
+export type AuthSetting = { credentials: Object, options: Object }
+export type LoginCommandOf<A extends AuthSetting, N extends string> = LoginCommand<N, A['credentials']>
+export type EncryptFunction<A extends AuthSetting, M> = (str: A["credentials"][Extract<keyof M[Extract<keyof M, string>], string> & Extract<keyof A["credentials"], string>], options?: Object) => string
