+++ modules/standards/src/standard-user-definition.ts
@@ -10,7 +10,6 @@ import { removePasswordFromResponseData } from './remove-password-from-response-
 import {
   GeneralReqResEntityMap,
   Key,
-  LoginCommand,
   EntityClient,
   EntityDefinition,
   UserDefinition,
@@ -22,19 +21,14 @@ import {
 
 import {
   EncryptFunction,
+  RestApiExecution,
+  AuthSetting,
+  LoginCommandOf
 } from '../decls/index'
-
-// @TODO: should we put those types into @phenyl/interfaces?
-type RestApiExecution = (
-  reqData: GeneralRequestData,
-  session: Session | null | undefined
-) => Promise<GeneralResponseData>
-export type AuthSetting = { credentials: Object, options: Object }
-export type LoginCommandOf<A extends AuthSetting, N extends string> = LoginCommand<N, A['credentials']>
 
 export type StandardUserDefinitionParams<M extends GeneralReqResEntityMap, A extends AuthSetting> = {
   entityClient: EntityClient<M>,
-  encrypt?: EncryptFunction,
+  encrypt?: EncryptFunction<A, M>,
   accountPropName?: Key<M[Key<M>]> & Key<A['credentials']>,
   passwordPropName?: Key<M[Key<M>]> & Key<A['credentials']>,
   ttl?: number,
@@ -42,7 +36,7 @@ export type StandardUserDefinitionParams<M extends GeneralReqResEntityMap, A ext
 
 export class StandardUserDefinition<M extends GeneralReqResEntityMap = GeneralReqResEntityMap, A extends AuthSetting = AuthSetting> extends StandardEntityDefinition implements EntityDefinition, UserDefinition {
   entityClient: EntityClient<M>
-  encrypt: EncryptFunction
+  encrypt: EncryptFunction<A, M>
   accountPropName: Key<M[Key<M>]> & Key<A['credentials']>
   passwordPropName: Key<M[Key<M>]> & Key<A['credentials']>
   ttl: number
@@ -51,12 +45,15 @@ export class StandardUserDefinition<M extends GeneralReqResEntityMap = GeneralRe
     super(params)
     this.entityClient = params.entityClient
     this.encrypt = params.encrypt || powerCrypt // TODO: pass salt string to powerCrypt
+    // @ts-ignore default
     this.accountPropName = params.accountPropName || 'account'
+    // @ts-ignore default
     this.passwordPropName = params.passwordPropName || 'password'
     this.ttl = params.ttl || 60 * 60 * 24 * 365 // one year
   }
 
-  async authentication<N extends Key<M>>(loginCommand: LoginCommandOf<A, N>, session: Session | null | undefined): Promise<AuthenticationResult<string, any>> { // eslint-disable-line no-unused-vars
+  // @TODO: fix this implmented
+  async authorize<N extends Key<M>>(loginCommand: LoginCommandOf<A, N>, session: Session | null | undefined): Promise<AuthenticationResult<string, any>> { // eslint-disable-line no-unused-vars
     const { accountPropName, passwordPropName, ttl } = this
     const { credentials, entityName } = loginCommand
 
@@ -80,6 +77,7 @@ export class StandardUserDefinition<M extends GeneralReqResEntityMap = GeneralRe
     }
   }
 
+  // @TODO: fix this implmented
   async wrapExecution(reqData: GeneralRequestData, session: Session | null | undefined, execution: RestApiExecution): Promise<GeneralResponseData> {
     const newReqData = encryptPasswordInRequestData(reqData, this.passwordPropName, this.encrypt)
     const resData = await execution(newReqData, session)
